backend/requirements.txt
fastapi 
uvicorn[standard] 
python-multipart 
opencv-python 
mediapipe 
numpy 
pillow 
requests 
python-dotenv 